requirements.txt
torch>=2.0.0
onnx>=1.13.0
numpy>=1.22.0
pytest>=7.0.0
torchvision>=0.15.0
timm>=0.9.0
transformers>=4.30.0
onnx2torch>=1.5.0
onnxscript>=0.2.0
onnx-simplifier>=0.4.0